(maint) Reduce travis notifications to a one-liner
This version drops the repo owner (all ours are puppetlabs so
no info there), the sha (do we care?), the author (there can be
several), and the change url, and collapses the build url onto
that one line.

It's an attempt at improving the s/n ratio of these notifications
to see if they're worth sticking with.