(PUP-1942) Initial pass at a global scope
The new Puppet::Pops::Evaluator::Scopes class is a facade over the
various scope types that will exist. The idea is that the scopes
themselves will be very dumb and fast, whereas the scopes class will
contain extra checks and convienence methods.

This is an very simplistic first pass and a global scope. that can set
and lookup values. Right now there isn't much to it, but it should grow
over time as the invariants and preconditions are discovered.