(PUP-2168) Fix systemd service enumeration on system 212
systemd 212 has a different output format for the `systemctl list-services`
command, which includes some white space with some circle symbols, before the
unit names. Without this patch, `puppet resource service` does not enumerate
any services on the system at all, running or otherwise.

Additionally, the "not-found" state seems to be possible as well as "error"
and "loaded", so include that as well.