(maint) Remove shell_out fallback code in yumhelper
The yumhelper had two paths of code - either it could load the yum
library and would use this to enumerate the updates, or it would call
`yum check-update` and parse the output.

The latter was typically used on very old versions of Yum (i.e. RHEL3
era) which have long been EOL.  To simplify this code, the shell
fallback method has been removed.  Failures to load the library will
result in the script raising an error.