(maint) Update external_ca_support/fixtures with 2048 bit keys
A recent security update is causing httpd package on RHEL6.5 to bring in
a more recent openssl (1.0.1e-16.el6_5.4) which was failing to work with
the 512bit RSA keys in our test fixtures.

This commit updates the certchain.sh to produce 2048bit RSA keys.  This
script was then run to produce a new set of fixture artifacts, and the
cert/keys for the fixtures were then updated from its output.