(PUP-2026) Add lambda/callable type to model
This adds a Callable type to the type system (the model), and makes
use of it in the new Function API (to return the type(s) of
a function's dispatcher - i.e. the supported signature(s).

A single Callable is returned for a single signature, and a Variant
for multiple signatures.