(PUP-2158) Unlist configured environment
The addition of a default configured environment when setting up
directory environment loaders without legacy environment loaders is
necessary to provide a bootstrapping env for puppet.  But this env was
also being listed when puppet was queried about it's enumerable
environments.  Since the static configured environment is not actually
useable in a meaningful way by an agent, it makes no sense to list it.
This commit provides a StaticPrivate loader, which, like Legacy loader
does not list its environments.