(maint) Beaker tests useful for reviewing environment status.
Currently these tests fail, but they do not fail fast so that all data
is collected and output in a review at the end of each test.  This let's
us take a look at the overall state of environment functionality in
varios configurations.

If we fix the failures, these could be moved into acceptance/tests as a
regression test.