(PUP-2176) Loading functions via new loaders should use code's loader
Without this change, loading functions is always done from the
perspective of the environment, not the perspective of the loader
that loaded the code that makes the call. This is an important feature
not only for modularity, but also for being able to conveniently
test with a test specific loader with test specific functions.

This change simply searches for a LoaderAdapter and favors that
loader over the environment loader. This mechanism will be used
for everything loaded, but is not yet implemented in the loaders).

This fix was on master, but was lost in a merge conflict.