(maint) Format according to norms
This fixes up a few formatting errors and converts the new (as well as
the old) uses of lambda {}.should to be the newer expect {}.to. It also
fixes a few tests that turned out to be making incorrect assertions. The
block to raise_error doesn't cause anything to fail if it returns false.