(PUP-2035) Adapt autoloaded functions with loader
All puppet code that is loaded needs to be associated with the loader
that loaded that code. This is important for subsequent requests to load
things from the body of the loaded code as it should use the same loader
that loaded that body of code.

The evaluator uses this information when calls are made.
Currently, if code lacks information about loader, the private
environment loader is used and it has visibility into all modules.

This commit takes a small step in the right direction of implementing
this everywhere as it adds a loader adapter to autoloaded puppet
functions.