(PUP-1799) Remove exposure of TypeFactory in Functions
This removes the access to all of the TypeFactory methods in the
Function interface. Users will have to use either the basic ruby types
(String, Integer, Fixnum, etc.) or use strings to specify the puppet
types ('Enum["a", "b"]', 'String', etc.). This reduces the API that
users will have to deal with and which we expose from the internals of
puppet.