(PUP-1799) Clarify public/private portions of the Functions API
Moved attribute/parameter injection support and docs to the tail end of the
file, and marked api private.  Also went through and designated the
public methods of the Functions API (create_function(), and the methods
such as dispatch(), param() and so forth which may be used to build up
dispatching for a function).