(PUP-2284) Add --skip-interview option to PMT generate.
Prior to this commit, we had a non-optional interview which helpfully
pre-populated common module metadata fields with values from the user.
However, with speculation of non-interactive uses of the `generate`
action (including our own acceptance test suite), this commit adds an
option enabling the user to bypass the interview entirely.