(maint) Remove an outdated PMT acceptance test.
Prior to this commit, we had an acceptance test in place to prevent the
PMT from accidentally breaking backwards compatibility with the Gem
version of the same tool.  Since that version is now long outdated, and
since the core-Puppet version of the tool has long since solidified its
own namespace in such a way that it does not conflict, continuing to
assert that we do not conflict is of limited value.