(maint) Clarify tests for loaders
As I went through the tests for loaders it seemed like there was some
overlap and some that were not clearly stating their purpose. This tries
to put clearer names on them and also shows where a test is missing:
what is visible to a module that *does* have metadata.