(PUP-2344) Remove dead code path
This was trying to make a decision on whether the environmentdir
setting is nil. Unfortunately it is always nil because it doesn't exist.
So this removes the dead path where it is not nil.

This code needs to be redone anyway because the environmentpath setting
is more complicated, and probably shouldn't be used here anyway.