(maint) Remove dependency on context system
I'd like to keep the pops code as independent of the other puppet
subsystems as possible. In this case the loaders were dependent on the
context system to get the overall loaders. Instead of relying on this
external system of global information, the loaders instance can just be
passed into the constructor instead. This also has the advantage of
removing a lot of context overrides that were being done in the tests.