(maint) Convert to using have_resource matcher
Using the have_resource matcher makes the tests a little shorter and
more expressive. It also uses the common way of matching so that we can
get better error reporting.

It turns out that one of the tests was matching against the incorrect
class name, so that is now fixed.

This also extracts a be_resource matcher so that you can match against
an individual resource independent of a catalog.